Rename dart to pub
