+++ internal/pub/pub_config.go
@@ -1,4 +1,4 @@
-package dart
+package pub
 
 import (
 	"encoding/json"
